+++ src/training/graph_group.h
@@ -13,7 +13,7 @@ namespace marian {
 
 // With -Ofast enabled gcc will fail to identify NaN or Inf. Safeguard here.
 static inline bool isFinite(float x) {
-#ifdef __GNUC__ 
+#ifdef __GNUC__
   ABORT_IF(std::isfinite(0.f / 0.f), "NaN detection unreliable. Disable -Ofast compiler option.");
 #endif
   return std::isfinite(x);
@@ -27,7 +27,7 @@ static inline bool isFinite(float x) {
 // if one value is nonfinite propagate Nan into the reduction.
 static inline void accNanOrNorm(float& lhs, float rhs) {
   if(isFinite(lhs) && isFinite(rhs)) {
-    lhs = sqrtf(lhs * lhs + rhs * rhs); 
+    lhs = sqrtf(lhs * lhs + rhs * rhs);
   } else
     lhs = std::numeric_limits<float>::quiet_NaN();
 }
@@ -42,39 +42,39 @@ static inline void accNanOrNorm(float& lhs, float rhs) {
 class GraphGroup {
 protected:
   Ptr<Options> options_;
-  
+
   Ptr<ICommunicator> comm_; // [not null] communicator, e.g. NCCLCommunicator
   Ptr<IMPIWrapper> mpi_;    // [not null] all MPI-like communication goes through this (this is a dummy implementation if no MPI run)
 
   std::vector<DeviceId> devices_;                   // [deviceIndex]
-  ShardingMode shardingMode_{ShardingMode::global}; // If local and multi-node training, shard only on local devices and do full sync (faster). If global shard across entire set of GPUs (more RAM). 
-  
+  ShardingMode shardingMode_{ShardingMode::global}; // If local and multi-node training, shard only on local devices and do full sync (faster). If global shard across entire set of GPUs (more RAM).
+
   // common for all graph groups, individual graph groups decide how to fill them
   std::vector<Ptr<ExpressionGraph>> graphs_;            // [deviceIndex]
   std::vector<Ptr<models::ICriterionFunction>> models_; // [deviceIndex]
   std::vector<Ptr<OptimizerBase>> optimizerShards_;     // [deviceIndex]
 
   Ptr<Scheduler> scheduler_; // scheduler that keeps track of how much has been processed
-  
+
   bool finalized_{false};    // 'true' if training has completed (further updates are no longer allowed)
   double typicalTrgBatchWords_{0}; // for dynamic batch sizing: typical batch size in words
   bool mbRoundUp_{true}; // round up batches for more efficient training but can make batch size less stable, disable with --mini-batch-round-up=false
 
-  bool costScale_{false};
-  float costScaleFactor_{1.f}; // @TODO, add current costScaleFactor_ to trainingState for serialization
-  size_t costScaleFreq_{2000};
-  float costScaleMultiplier_{2.f};
-  float costScaleNanTolerance_{0.f};
-  size_t costScaleNanRange_{1};
-  float costScaleFactorMinimum_{1.f}; // @TODO make this configureable
+  bool costScaling_{false};
+  float costScalingFactor_{1.f}; // @TODO, add current costScalingFactor_ to trainingState for serialization
+  size_t costScalingFreq_{2000};
+  float costScalingMultiplier_{2.f};
+  float costScalingFactorMinimum_{1.f};
+
   size_t noNanSeen_{0}; // @TODO, add current noNanSeen_ to trainingState for serialization
   size_t nanSeen_{0};
+
+  bool checkGradientNan_{false};
 
   bool dynamicGradientScaling_{false};
   float dynamicGradientScalingFactor_{2.f};
   bool dynamicGradientScalingUseLogs_{false};
-
-  bool checkGradientNan_{false};
+  size_t dynamicGradientScalingFadeout_{0ul};
 
   // determines the number of input streams (i.e. input files or fields in the TSV input) that need
   // to be included in the batch, i.e. without alignments and weights
@@ -100,16 +100,16 @@ public:
 
   virtual void load();
   virtual void save(bool isFinal = false);
-  
+
 private:
   void load(const OptimizerBase::ScatterStateFunc& scatterFn);
   void save(bool isFinal,
             const OptimizerBase::GatherStateFunc& gatherOptimizerStateFn);
 
-  bool restoreFromCheckpoint(const std::string& modelFileName, 
+  bool restoreFromCheckpoint(const std::string& modelFileName,
                              const OptimizerBase::ScatterStateFunc& scatterFn);
 
-  void saveCheckpoint(const std::string& modelFileName, 
+  void saveCheckpoint(const std::string& modelFileName,
                       const OptimizerBase::GatherStateFunc& gatherFn);
 
 public:
@@ -128,11 +128,11 @@ public:
   float executeAndCollectNorm(const std::function<float(size_t, size_t, size_t)>& task);
 
   float computeNormalizationFactor(float gNorm, size_t updateTrgWords);
-  
+
   /**
    * Determine maximal batch size that can fit into the given workspace
    * so that reallocation does not happen. Rather adjust the batch size
-   * based on the stastistics collected here. Activated with
+   * based on the statistics collected here. Activated with
    * `--mini-batch-fit`.
    * In a multi-GPU scenario, the first GPU is used to determine the size.
    * The actual allowed size is then determined by multiplying it with the